define.scala
